% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df1.R
\docType{data}
\name{df1}
\alias{df1}
\title{Example dataset with longitudinal counts}
\format{
A data frame with 18 rows and 5 variables
}
\usage{
data(df1)
}
\description{
An example dataset of a study with longitudinal counts, used
to illustrate how \code{ptmixed()} and \code{nbmixed()} work.
}
\references{
Signorelli, M., Spitali, P., Tsonaka, R. (2020). Poisson-Tweedie
mixed-effects model: a flexible approach for the analysis of longitudinal RNA-seq
data. Statistical Modelling. URL: https://doi.org/10.1177/1471082X20936017
}
\seealso{
examples in the \code{\link{nbmixed}} and
\code{\link{ptmixed}} help pages
}
\author{
Mirko Signorelli
}
\keyword{datasets}
