% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-ptglm.R
\name{summary.ptglm}
\alias{summary.ptglm}
\title{Summarizing Poisson-Tweedie GLM estimation results}
\usage{
\method{summary}{ptglm}(object, ...)
}
\arguments{
\item{object}{an object of class \code{ptglm} (obtained from \code{ptglm}).}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A list with the following elements: \code{logl}, \code{coefficients}, 
\code{D}, \code{a}
}
\description{
Provides parameter estimates, standard errors and univariate Wald test
for the Poisson-Tweedie generalized linear model
}
\seealso{
\code{\link{ptglm}} and the examples therein
}
\author{
Mirko Signorelli
}
