% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df1.R
\docType{data}
\name{df1}
\alias{df1}
\title{Example dataset with longitudinal counts}
\format{A data frame with 18 rows and 5 variables}
\usage{
data(df1)
}
\description{
An example dataset of a study with longitudinal counts, used
to illustrate how `ptmixed()` works.
}
\author{
Mirko Signorelli
}
\keyword{datasets}
