% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.ptglmm}
\alias{summary.ptglmm}
\title{Summarizing Poisson-Tweedie mixed model estimation results}
\usage{
\method{summary}{ptglmm}(object, wald = T, ...)
}
\arguments{
\item{object}{an object of class \code{ptglmm} (obtained from \code{ptmixed}).}

\item{wald}{logical. If \code{TRUE}, standard errors and univariate Wald test are computed. Default is \code{TRUE}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A list with the following elements: \code{logl}, \code{coefficients}, 
\code{D}, \code{a}, \code{sigma2}
}
\description{
Provides parameter estimates, standard errors and univariate Wald test
for the Poisson-Tweedie generalized linear mixed model
}
\seealso{
\code{\link{ptmixed}} and the examples therein
}
\author{
Mirko Signorelli
}
